# swagger.yaml
swagger: '2.0'
info:
  description: Arthur Clive Web Api
  version: 1.0.0
  title: Arthur Clive Swagger
  contact:
    email: raguvarthan.n@turingminds.com
host: arhturclive.swagger.io
basePath: /v1
...
schemes:
  - http
paths:
...
  '/api/category/{productFor}/{productType}':
    delete:
      tags:
        - CategoryController
      operationId: deleteCategory
      produces:
        - application/xml
        - application/json
      parameters:
        - name: productFor
          in: path
          description: productFor of the category that needs to be deleted
          required: true
          type: string
        - name: productType
          in: path
          description: productType of the category that needs to be deleted
          required: true
          type: string
      responses:
        '200':
          description: Deleted
          schema:
            $ref: '#/definitions/ResponseData'
        '400':
          description: Failed
          schema:
            $ref: '#/definitions/ResponseData'
        '404':
          description: Product Not Found
          schema:
            $ref: '#/definitions/ResponseData'
  '/api/subcategory/{productFor}/{productType}':
    get:
      tags:
        - SubCategoryController
      operationId: getSubCategories
      produces:
        - application/xml
        - application/json
      parameters:
        - name: productFor
          in: path
          description: productFor of the subcategory that needs to be filtered
          required: true
          type: string
        - name: productType
          in: path
          description: productType of the subcategory that needs to be filtered
          required: true
          type: string
      responses:
        '200':
          description: Success
          schema:
            $ref: '#/definitions/ProductDetails'
        '400':
          description: Failed
          schema:
            $ref: '#/definitions/ResponseData'
...
  '/api/admin/subscribe/{username}':
    post:
      tags:
        - AdminContoller
      operationId: subscribe
      consumes:
        - application/json
        - application/xml
      produces:
        - application/xml
        - application/json
      parameters:
        - in: path
          name: username
          description: UserName of user who needs to subscribe
          required: true
          type: string
      responses:
        '200':
          description: Subscribed Succesfully
          schema:
            $ref: '#/definitions/ResponseData'
        '400':
          description: Failed
          schema:
            $ref: '#/definitions/ResponseData'
        '401':
          description: User Already Subscribed
          schema:
            $ref: '#/definitions/ResponseData'
        '402':
          description: UserName connot be empty
          schema:
            $ref: '#/definitions/ResponseData'
  '/api/admin/unsubscribe/{username}':
    post:
      tags:
        - AdminContoller
      operationId: unsubscribe
      consumes:
        - application/json
        - application/xml
      produces:
        - application/xml
        - application/json
      parameters:
        - in: path
          name: username
          description: UserName of user who needs to unsubscribe
          required: true
          type: string
      responses:
        '200':
          description: Unsubscribed Succesfully
          schema:
            $ref: '#/definitions/ResponseData'
        '400':
          description: Failed
          schema:
            $ref: '#/definitions/ResponseData'
        '401':
          description: No user found
          schema:
            $ref: '#/definitions/ResponseData'
        '402':
          description: UserName connot be empty
          schema:
            $ref: '#/definitions/ResponseData'
  /api/admin/sendmessage:
    post:
      tags:
        - AdminContoller
      operationId: publicPost
      consumes:
        - application/json
        - application/xml
      produces:
        - application/xml
        - application/json
      parameters:
        - in: body
          name: message
          description: Message that needs to be sent to the subscribed users
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Email sent to all subscribed users
          schema:
            $ref: '#/definitions/ResponseData'
        '400':
          description: Failed
          schema:
            $ref: '#/definitions/ResponseData'
        '401':
          description: There are no subscribed users
          schema:
            $ref: '#/definitions/ResponseData'
...
definitions:
  ResponseData:
    type: object
    properties:
      Code:
        type: number
      Message:
        type: string
      Data:
        type: object
      Content:
        type: object
...
  ProductDetails:
    type: object
    properties:
      Id:
        type: number
      ProductSKU:
        type: string
      MinioObject_URl:
        type: string
      ProductFor:
        type: string
      ProductType:
        type: string
      ProductDesign:
        type: string
      ProductBrand:
        type: string
      ProductPrice:
        type: number
      ProductDiscount:
        type: number
      ProductDiscountPrice:
        type: string
      ProductStock:
        type: number
      ProductSize:
        type: string
      ProductMaterial:
        type: string
      ProductRating:
        type: number
      ProductReviews:
        $ref: '#/definitions/Reviews'
      ProductColour:
        type: string
      RefundApplicable:
        type: boolean
      ReplacementApplicable:
        type: boolean
      ProductDescription:
        type: string
  Reviews:
    type: object
    properties:
      Name:
        type: string
      Comment:
        type: string
...